Extract the move machinery, and ignore the flat root
`Move`, `applyMoves`, `assertOneObjectPerDestination` and `claim` were inside
the contract-3 migration; contract 4 relocates bytes too and needs the same
rules — nothing deleted until what replaced it has been read back, every
destination checked before the first byte moves, a rerun after an interruption
finishing rather than refusing. Each of those is a bug this project already
paid for once, so they live in one module rather than being reimplemented
beside a second migration.

`contract3-data.ts` drops from 399 lines to 262 as a result.

`.gitignore` follows the flat root: `profiles/` whole rather than a glob over
its files, since a profile is a directory now, plus the four workspace stores
`data/` used to cover. This repo doubles as a workspace during local dev, so
the file is the first line of defence rather than housekeeping.
