+++ src/cli/contract3.test.ts
@@ -5,7 +5,8 @@ import { join } from 'node:path';
 import { parse } from 'yaml';
 import { migrateToContract3 } from './contract3.ts';
 import { migrateToCurrentContract } from './workspace-migrate.ts';
-import { applyMoves, planMoves } from './contract3-data.ts';
+import { planMoves } from './contract3-data.ts';
+import { applyMoves } from './migrate-move.ts';
 import { planCredentials } from './contract3-credentials.ts';
 import { createMemoryBlobStore } from '#stores/blobs/testing.ts';
 import { createFileSecretStore } from '#secrets';
